+++ RCISystem/DAL/Interface/IAccountantsAdviceDAL.cs
@@ -14,7 +14,7 @@ namespace RCISystem.DAL
         //bool UpdateBank(BankModel bankModel);
         //bool DeleteBank(int Id);
         //DataRow GetBankById(int Id);
-        DataTable SearchByAccountantsAdviceID(int Id);
-        int AddAccountantsAdviceTransaction(AccountantsAdviceModel accountantsAdviceModel,List<AccountantsAdviceDetailModel> accountantsAdviceDetailModel);
+        DataTable SearchByAccountantsAdviceID(decimal Id);
+        decimal AddAccountantsAdviceTransaction(AccountantsAdviceModel accountantsAdviceModel, List<AccountantsAdviceDetailModel> accountantsAdviceDetailModel);
     }
 }
