+++ RCISystem/DAL/Interface/ICheckIssuedDAL.cs
@@ -11,7 +11,7 @@ namespace RCISystem.DAL
     {
         DataTable GetAllCheckIssued(SearchField searchField);
         bool AddCheckIssued(CheckIssuedModel checkIssuedModel);
-        int AddCheckIssuedTransaction(CheckModel checkModel,List<VoucherModel> voucherModel,int UserID);
+        decimal AddCheckIssuedTransaction(CheckModel checkModel, List<VoucherModel> voucherModel, int UserID);
         //int AddCheckIssuedTransaction(CheckModel checkModel, string[,] voucherInfo);
         //bool UpdateBank(BankModel bankModel);
         //bool DeleteBank(int Id);
